requirements.txt
asgiref==3.4.1
Django==3.0.7
django-filter==2.4.0
future==0.18.2
html5lib==1.1
Pillow==8.3.2
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.1
reportlab==3.6.1
six==1.16.0
sqlparse==0.4.2
webencodings==0.5.1
xhtml2pdf==0.2.5